Parsing out code to create adaptable functions
Parsing code for conversion of elevation change to melt rate so that functions are called to grab RACMO and firn air content data (as needed) and iterate for density estimation. Adding prompts to ask about study site: Greenland, Antarctic Peninsula, or Antarctica. Sites are needed for RACMO file identification and projection of data.