modified DEM date extraction from PGC metadata
DEMs produced by the Polar Geospatial Center have different metadata saved in their text files depending on the version of the algorithm (SETSM) used to generate the DEMs. If DEMs were produced with SETSM 3.1.0, and possibly other versions, the acquisition time was not saved as a separate variable and needs to be extracted from input image file names.